This is a big update. Here are the highlights :
The client server bridge now has an additional guest_token that
tracks objects handed out to visitors that are not logged in. The
system gives a token and notes the ip address to map the token
back to the originator.

Yote::Obj has been organized. Non API methods all now start with
an underscore ( except for new )